fix: set CAM field properly in convenience methods
Currently, the convenience methods for creating ACK packets (exec,
validation, and query ACKs) have a bug that causes a parsing error on
the other end:

```
CAM field in ACK packet does not select one of Validation, Exec, or Query
```

The issue is that when creating these exec packets, we set the default
CAM inside of our enum wrapper instead of explicitly setting the ACK type.
This is subtle as it allows the ACK sender to disambiguate the ACK type,
but not the receiver.

Fix this bug, add test cases that catch it, and improve the error
message a little bit to make it clear that only one of those types of
ACKs needs to be set.

Signed-off-by: John Moon <john.moon@voyagertechnologies.com>